Vim: turn off delimitMate inside String groups
When writing Rust code, string interpolation is done by putting a pair
of braces ({}) inside the string and then adding something to be
evaluated in the following arguments.

If delimitMate_jump_expansion is turned on then attempting to add the
closing brace inside the string (for the purpose of interpolation) will
result in the cursor jumping to the position of an already existing
closing brace (for instance, the brace closing the function).

While I'm currently not sure if this change will significantly affect
coding in other languages, I think it should be fine.