Vim: Replace my own fork of Mark with offical one
About two years ago I forked this repository from the (still)
out-of-sync repository found on the vim-scripts account.

It started out well, updating the plugin to the (at the time) latest
version, but then I started hard-coding mappings and settings because
I didn't read the code around the lines that had "nmap" on them.

Apparently I was both lazy and silly back then.

This patch changes the submodule to point to a repository that updates
to the latest version available on vimscripts.org.