Vim: remove auto-nohl for searches
This reverts the changes introduced in bc1184b22 that were related to
automatic nohl. Those changes also prompted a change in the mappings for
the Mark plugin, so revert those too.

I found that I'm too used to doing the nohl with double Escape (or
rather double Tab, since I've bound it that way). Having highlighting
disappear when I don't want it to was irritating enough that it was
unacceptable.