Git: fix glitchy commit hook
The pre-commit hook is meant to disallow committing lines that contain
the keywords defined in the same file.

Currently, if there are no files changed in the repository and the
repository *in general* contains one or more of the keywords, then these
will be found by the `git grep`-line in the pre-commit hook. This is
caused by the fact that `git grep` without any file arguments (joining
an empty list of filenames is still an empty string) results in the
entire repository being searched.

This results in a non-zero exit status for the `git grep` command, and
subsequently the pre-commit hook, preventing a commit for no reason.

This patch makes the hook behave sanely.