Vim: disable vim-racer if no Rust source
The vim-racer plugin doesn't really have a feature which I could combine
with rcm's host-based configuration, to turn the plugin off on computers
that don't have the Rust source installed into the $RUST_SRC_PATH
directory.

Setting the flag for whether the plugin is loaded prevents the rest of
the plugin source to be evaluated.