Zsh: add cabal binary directory to PATH
While the module provided by Prezto adds the ~/.cabal/bin directory
properly to the PATH (it's currently the only thing it does), since it's
a Prezto module, it won't get loaded in non-interactive shells.

This means that if something would need binaries from ~/.cabal/bin in
a non-interactive shell, it won't find it despite having the module
loaded.