Vim: For Haskell, indent where-lines by 6
I'll try to go with the defaults and indent lines after a "where"
keyword to the width of "where" plus one space. Right now it feels a bit
weird, but I might come to like it.

Besides, it seems like the haskell-vim plugin isn't configured to have
a two-space indented "where" on its own line: the plugin expects to have
code following the "where".