Neovim: change terminal escape mapping
In Neovim, the fzf plugin spawns fzf in a Terminal buffer and has Esc as
a mapping for aborting fzf and closing the buffer. Currently Esc is
mapped to turning off Insert-mode in a Terminal buffer, meaning that Esc
stops the input to fzf, but doesn't actually close the Terminal buffer.

Since I use Terminal buffers quite rarely, change the Esc mapping to
<C-\> (shorter version of the original mapping).