i3: Fix occasionally broken temp monitor in bar
On primo and gyllenborste it seems like the folder
`/sys/devices/platform/coretemp.0/hwmon/hwmonX/temp1_input` changes its
X to different integer values, most often to either 1 or 2. Every time
this makes the path differ from the path specified in the configuration
file for the status bar, the temperature module breaks.

Since it doesn't seem like it matters what number the folder has (and
there's never more than one folder), find the path with the right number
and use it in the temperature module.