Vim: stow away rarely used leader mappings
I feel like I'm starting to run out of <leader>X mappings, where X is
any letter. While there are technically a bunch of letters available,
I think that it would be good to stow away mappings that are not very
often used under some grouping key, to namespace them.

Therefore, put a couple of mappings under the <leader>a namespace (right
now the mnemonic is the Swedish word "annat" or "all the others").