Neovim: turn off vim-go text objects
While it might be handy to have a text object for functions in go, the
default binding for those objects is `if` in visual mode. This clashes
with my mapping for moving one character to the left.

Since I don't feel like digging into this right now, disable the text
objects for now.