Neovim: update whitespace highlighting mappings
In a recent commit I "enabled" these mappings once more by making them
use the same leader key as all the other mappings are. Since the time
I changed the leader key I've added the <leader>w mapping for writing
the current buffer, which "conflicts" with the highlighting mappings (in
the sense that the <leader>w mapping is no longer instant).

Put the highlighting mappings under the <leader>a namespace to resolve
the conflict.