Neovim: fix helper mapping overwriting search
The line splitting mapping makes use of substitution to remove potential
trailing whitespace when applied. When defined as a mapping, this
overwrites any previous search that existed.

By placing the search (and the rest of the procedure) inside a function,
the previous search will not be overwritten.

The `exe` command in this patch is used because I couldn't figure out
how to get the `<CR>` to be interpreted as one character when just using
`norm!`.