Neovim: remove redundant options
Neovim has set defaults for certain options that mirror my
configuration, making the inclusion of them in my `init.vim` redundant.

In the case of `history` the default is 10000, which is something
I don't mind bumping up to.

As for the `tags` option, my current setting is identical to the
default, apart from me including capitalized versions of the word
"tags". Since I have yet to come across tags-files that are named like
that, I'll go with the default that only looks for the lowercase version
of the word.