Neovim: use upstream vim-ruby plugin
A long time ago I made my own fork of vim-ruby to address the following
issues I had with the plugin:

* Disable some mappings that didn't mesh well with colemak.vim
* Prevent some automated alignment of tokens

Upstream allows for easy disabling of mappings, so this is not a problem
anymore. As for the alignment of tokens, some brief usage of the
upstream plugin gives me the impression that this not a problem either.

Change the vim-ruby source to upstream.