todo.sh: replace `mit` add-on with my own rewrite
While the `mit` add-on is tremendously helpful, I've had various issues
with it that I didn't feel like wrestling with by digging into its code.

My rewrite is almost a drop-in replacement and improves on a couple on
areas, such as how MIT dates are added and removed to TODOs.