Revert "Neovim: remove incsearch plugin"
This reverts commit d878057a84e5e0f096e2b928ff540c1980ffa3e6.

The incsearch plugin was removed because it broke when using the
development version (latest commit on master) of Neovim. After getting
fed up with how much things were breaking, I opted to go back to the
latest stable version of Neovim. This means that the incsearch can be
added back without issue.