Neovim: change completeopt setting
The `longest` option makes sense if one only brings up the completion
menu every now and then, but since I use deoplete, the completion is
invoked pretty much every keystroke. This results in the following
frustrating behaviour:

If I was to, for instance, have the cursor placed at the end of a word
with another occurrence and then rapidly (and repeatedly) press <BS>,
the word would not be removed; as a result of the word being the
`longest` available every time deoplete makes the completion menu pop
up, it gets replaced to its full length.

Change `completeopt` to use `noinsert` instead of `longest`.