i3: add extra tabbed layout mapping on kvaser
kvaser will generally have a Kinesis Advantage plugged into it. This
keyboard has its [ and ] keys moved to the lower right part of the right
hand half of the keyboard.

To avoid messing too much with muscle memory, add a duplicate mapping on
the backslash key, which is positioned in the same space that [ would
normally be.