mpd: use explicit address for host
I don't know if it's because one of the latest versions of mpd has
changed something but on kvaser it seems like mpd is more sensitive to
the problem related to binding IPv6 before IPv4, resulting in this error
on startup:

  listen: bind to '0.0.0.0:6600' failed: Address already in use (continuing anyway, because binding to '[::]:6600' succeeded)

Solution found here:

  https://wiki.archlinux.org/index.php/Music_Player_Daemon/Troubleshooting#Binding_to_IPV6_before_IPV4

It could be this, or it could be the fact that the domain name
"localhost" is bound to something else than 127.0.0.1 in the office
where kvaser resides. Either way, explicitly stating the host address
seems like a good thing to do anyway.