mc: use provided skin instead of defining colors
Some of the provided skins look quite nice and the only thing my list of
Colors did was to make those skins look terrible, while doing a worse
job than the chosen skin.

Also include the editor_state_full_filename and panel_scroll_center
settings, which seem to be new things in the config file that mc insist
on including.