todo.sh: move out of SpiderOak
SpiderOak sometimes syncs files in ways that result in unexpected
changes to quickly changing files, like todo.txt can be when one
adds/completes a bunch of tasks. It gets extra frustrating when one has
multiple computers that are all connected to the same SpiderOak account.

Having used the git add-on for todo.sh (syncing against a computer that
is not connected to SpiderOak), has been the only way to keep the
todo.txt contents reliable, and it has helped a lot with conflict
resolution.

Since it has worked so well, there's no reason for me to keep the
contents in SpiderOak.