NeoVim: disable theme transparency
By having transparency turned on for my theme, if I change the terminal
to have a different color palette (dark on light, instead of vice versa)
then having the transparency of the theme makes the background of the
editor be whatever the terminal background is, which does not
necessarily work out well.

Additionally, once upon a time I thought terminal transparency was
pleasing, but I no longer have my terminals do that.