Neovim: replace deoplete with coc.vim
I've generally been happy with deoplete, but the LanguageClient-neovim
plugin has been problematic at times. While I could have looked into
whether the vim-lsp plugin would have just worked better, I decided to
try out coc.vim for the purpose of improving my language server
completion experience.

coc.vim does indeed seem to work better than LanguageClient-neovim, but
is a complete replacement for everything that's related to completion.

Additionally, with LanguageClient-neovim + deoplete the language server
provided snippets broke and were mostly a pain to have suggested. With
the snippets actually working, this leads to it being reasonable to have
those go through coc.vim as well, which is why UltiSnips is removed from
my plugins (although the snippet collection from honza is kept).

For the completions of coc.vim (snippets or otherwise) to work properly,
the normal-mode <CR> mapping must be set to something specific. This
conflicts with the plugins delimitMate and endwise, which both remap
<CR> in some kind of passthrough fashion (delimitMate has to be
configured to do this, but that's the case in my config). This leads to
the removal of endwise, as I think I can do without it if I simply rely
more on snippets. Same goes for delimitMate: hopefully snippets and
automatic code formatting will be an acceptable replacement.