Neovim: add mapping for expanding snippets
Most of the time the completion window will open and let me choose
a snippet by pressing <CR>, but some snippets seem not to trigger the
window opening. Specifically, "#!" should result in showing the shebang
snippet, but seems not to.

While it could be that I should look into why the completion window is
not triggered instead of adding a mapping, I've worked up a muscle
memory for some things (the shebang line snippet, being one of them) so
this works fine for me.