Zsh: make less exit if less than one screen
I used to like always getting the screen cleared for `less` and not
exiting if the output is less than one screen. The main reason for this
was that I could be sure that if I read something with `less`, then
I would need to quit less (consistency is nice).

However, using value for `LESS` leads to various `git` commands clearing
the screen and locking me inside less, even if the output is one line.
This is often not ideal.

I suspect that I could somehow make the pager for `git` use a different
`LESS` value, but I don't feel like figuring this out at the moment.