Zsh: add `task` configuration and disable `todo`
The amount of time I've invested into tweaking `todo.sh` with plugins
comes as a consequence of finding out how I want to work with my TODO
manager.

Taskwarrior could be a better platform for me to work with, as
a consequence of its more structured way to store data.

The transition means I'll lose some of the FZF helper functions that
I've created over time, but these will probably be regained over time.