Zsh: run `task` aliases in subshells
Currently the functions in `tag-task/zsh_task_alias` use `pushd`/`popd`
to switch directories. The problem with this is that sometimes there are
unwanted directories on the stack. The objective is to carry out some
commands in the `$task_dir`, not actually move there: I never want to
move from the place where I ran one of the functions.

By wrapping the function definitions in a par of parentheses, a regular
`cd` can be used, as the actions are ran in subshells.