Zsh: fix broken `cd` in `task` aliases
In the cases where `task` data is broken, then `tcd` will be broken as
well.

Currently I ask the `task` command for what the `data.location` value
is, but this will not work if the data is in an abnormal state. The most
common case for me is when I get a merge conflict in the data files due
to fetching conflicting changes with git.

Instead of asking `task` for the data location, derive the `$task_dir`
from the `~/.taskrc` file.