Zsh: make `tlf`/`tlaf` filter by a fingerprint
The `description` for the various tasks is not always a sufficiently
granular piece of data for determining uniqueness.

I want to remove duplicates when filtering, but the list of tags a task
has can be an important differentiator, especially when `tlf`/`tlaf` is
used by the `tlfuuid`/`tlafuuid` functions.

Use a generated "fingerprint" key to filter for uniqueness.