Zsh: Set PASSWORD_STORE_DIR to .password-store
SpiderOak has once again shown to synchronize the various computers
unevenly: for some computers I need to do a manual restore for certain
files to get the latest version. Most times this is not an issue, but
for something like my password-store, this is unacceptable.

Since I'm already using automatic git-committing with `pass`, migrating
this to be a manually managed is fairly straight-forward.