Neovim: add Vimwiki function for generating tags
While I wanted a mapping for generating the tags, I've noticed that just
running `:VimwikiGenerateTags` does not seem to be enough. If the
`.tags` file is stale then the generated list becomes outdated, and
running `:VimwikiRebuildTags` has (IIRC) at times not resulted in
a rebuild according to expectations.

Additionally, the lines that `:VimwikiGenerateTags` generates are for
some reason generated with in the format of `[[file_name#Title]]`, which
will include the `file_name` in the rendered link name. I'm not sure if
this is a mistake or not, but I'll simply "patch" this by replacing the
`#` in the generated links with `|`.