Neovim: add `ferret` plugin
While `ferret` has a lot of overlap with the `grepper` plugin, the
quickfix window editing (specifically the removal of lines) is a feature
I really want to try out for a while. The other features of `ferret` may
end up being useful in some way as well.

With the introduction of this plugin I'll have `ferret`, `grepper` and
`CocSearch` from `coc.nvim` to help me with searching, so it may be that
one or more of these may need to go at some point.