Zsh: change `nvmail` to use `set tw=0`
It used to be the case that most of the email I wrote was wrapped at 72
characters, which was enforced by using `setf mail`. Occasionally I used
`nvmail` for quick writing that was not intended to be wrapped at 72
characters, like composing a GitHub comment with `nvmail`.

These days I find that I most often use `nvmail` for doing the latter,
i.e. various kinds of text that I just want write in NeoVim, for which
text wrapping does not make sense.

Additionally, I've come to the point where I think I've generally given
up on plain text email, which is the only place where hard wrapping of
the text makes sense (I'm excluding code related text, so this does not
apply to git commit messages, for instance). Most emails are read
in mail clients that render rich text better than plain text (GMail, for
instance), and if the mail client is on a mobile device then it's likely
that the plain text might get broken halfway through the line, resulting
in very painful reading. Letting the renderer handle the line wrapping
is simply the better option, I'm pretty sure.

Drop the hard wrapping as a default.