Neovim: remove explicit loading of matchit plugin
IIRC, I added the `matchit` plugin because the built-in version of the
plugin which came with Vim (which I used at the time) was not up-to-date
and some problems were arising.

It seems like now the reverse is happening: having the plugin included
seems to cause unexpected cases where pressing `%` simply does not do
what it should (no movement of the cursor).

Remove the plugin, as using the built-in seems to fix the problem.