i3: remove for_window lines from config
Once upon a time I played FTL and this line was a solution I found for
preventing the game from going to fullscreen immediately. Similarly,
`Gnome-screenshot` was an application that shouldn't get fullscreen'd
right away.

Since I no longer play FTL, and `Gnome-screenshot` does not seem to need
this type of line anymore to be shown as a floating window
automatically, I think it's time to remove these lines.

The main benefit of removing the lines is that the `~/.config/i3/i3log`
file will no longer be spammed to hell with sometimes gigabytes of log
output, mainly consisting of messages related to matching for these two
lines.