Neovim: replace my colorscheme with PaperColor
I don't develop my colorscheme, and I don't benefit from some upstream
developer improving the functionality of the colorscheme. Additionally,
my colorscheme seems to not work with `goneovim`, in case I would ever
decide to use it.

With PaperColor, I can get the colors I want by just applying a bunch of
overrides in the configuration. Another important detail is that this
theme, in contrast to some others, is made for dark/light background
switching, where a lot of themes seem to focus on one or the other.