NeoVim: use simpler trailing whitespace highlighting
I just learned that `listchars` has a `trail` option, which means that
I can highlight trailing whitespace in a rather easy fashion.
Additionally, it seems like this method will be slightly less noisy than
the red background that I've currently got.

It may be that I actually prefer to have the red background, but this
change offers me the opportunity to delete code and come a bit closer to
default NeoVim configuration, so I'll try this out.