Prezto: extract only the used from `utility` module
I had an idea that I should move to use `exa` more instead of `ls`,
which lead me to realize that I have a lot of configuration loaded from
the `utility` module of Prezto that I generally do not use.

Extract that which I actually use, then drop the usage of the module.
I guess this would also facilitate a future move away from Prezto, if
I decided to do that.