Neovim: change plugin_mappings.lua formatting
Changes that should have been made in e9a529597.

To make it easier to add Lua configuration for plugins, have each plugin
configuration be surrounded with the `vim.cmd` invocation. It's extra
noise, but having the organizational comments with Lua syntax and the
Vimscript configuration wrapped results in easier addition/deletion for
configuration.