Neovim: remove mapping of Tab to Escape
Big change, having had this configuration for years: I no longer want
Tab to behave like Escape in various modes.

This was an interesting idea added by Shai Coleman in the original
`colemak.vim`, with the intention of adding an Escape key closer to
one's fingers, without resorting to system level configuration. I still
agree with the general idea: Escape on the average keyboard is located
too far away for repeated usage. However, on my UHK I've got Escape
*much* more accessible (thumb), and on various other keyboards I could
probably remap a key like Caps Lock to act as Escape.

Overall, I've grown tired of reconfiguring various plugins to be
compatible with the fact that I don't have Tab available for them.

This change will be absolute hell for my muscle memory, but I'm thinking
that this is better in the long run.