Neovim: change mappings for leap.nvim
Since the rest of my movement mappings revolve around the idea of using
UNEI, LY, and JH, having `leap.nvim` do the more standard
lowercase/uppercase letter for a motion, it does not feel right.

Additionally, drop the usage of the `-x` functions (`leap-forward-x` and
`leap-backward-x`). I tried these out and really didn't feel like it
something I would remember, or want to block mapping spots for.

Finally, change the Grepper mappings to revolve around `gr` instead of
`gl`. This gives room for Leap, and also seems to be a better mnemonic.