Neovim: add mapping for :nohlsearch
Until recently (4decd34b0) I had Tab mapped to Escape, and beyond just
working as Escape, `:nohlsearch` was input after the Escape. This gave
me an easy way to clear the search highlighting.

After going from Tab to a normal Escape, I immediately noticed that
actually writing `:noh` was not going cut it. I don't know if the
mapping added in this patch is the final one, but it's at least
something.