Neovim: replace native LSP mappings with Telescope equivalents
I noticed that `vim.lsp.buf.references` loads up a quickfix list with
the references, but in contrast to coc.nvim, there is no preview for
those references. Use the Telescope equivalent instead, for the preview
functionality.

I'm assuming that `vim.lsp.buf.implementation` works the same way as
`references` (although I've yet to actually use it), so I'm switching it
to Telescope as well.