Neovim: change how `cmp` sources are loaded
The README for `cmp` has the sources added as two arguments to the
`sources` function. I collapsed the lists into one list, effectively
only calling the function with only one argument, as I thought that
there was some mistake in the README.

Having noticed that sometimes the completion doubles up some items in
the completion, I'm presuming that this might be the cause. It could be
that the second argument to the `sources` function might be fallback
sources, in case that the first list of sources do not produce any
results.