Neovim: turn off adding of mappings for `context.vim`
`context.vim` by default adds a mapping that overlaps with my remapped
undo (it maps `zt` when I use `z` for undo). This causes a delay upon
attempting to undo something, which is really annoying.

From reading the README for the plugin, it seems like there should be no
problem with turning off adding mappings (or more specifically, turning
off and not adding these mappings in some other place), because this
seems to be a feature added for old versions of Vim.