Neovim: upgrade and reconfigure vim-mark
I noticed that for some reason the `mark` plugin had one of the first
marks (4th, I think) get the same color as the one following it (5th).

I could have just tweaked the `mwDefaultHighlightingPalette` variable,
but instead I got the strange idea that the problem was that the plugin
was not compatible with my current set of plugins and current version of
Neovim. I don't know how I got this idea, but it lead me to upgrade
`mark`, by switching to the repository of the actual maintainer of the
plugin. This change of repository has resulted in moving from a repo
that hasn't been updated for the past 6 years, to one that was last
updated half a year ago.

Upgrading the plugin, however, didn't fix the original problem at all
(obviously, in hindsight), so I tweaked the
`mwDefaultHighlightingPalette` variable to have a list of colors that
work well with my current setup. The resulting list is much shorter, so
I'll have to expand it if I happen to more than 13 different things, at
some point.