Neovim: replace bullets.vim with autolist.nvim
I want to have the ability to write lists in Markdown as easily as
I currently do with Obsidian. Hitting <CR> at the end of a list item
should start a new list item at the same indentation level, and <CR> on
an empty bullet point should delete the bullet and give me a new line.
Basic stuff, but without a plugin this is just not the case.

In 6bce58ac I introduced bullets.vim, but noticed that there were
problems with how the plugin worked when one let the plugin apply the
default mappings (e.g. overriding <CR>). I don't remember what the exact
problem was, but it was sufficiently disruptive that I ended up
disabling all mappings and only use the `:RenumberList` command.

I haven't tested autolist.nvim too much, but the defaults seem to work
decently for me, so I'll try to see if this plugin works better
out-of-the-box (or with a little bit of configuration) than bullets.vim.