Neovim: remove `vim-markdown` and disable `markdown` for `vim-polyglot`
I've noticed that recently the `autolist.nvim` plugin started to add
extra list markers when pressing `<CR>` at the end of a line. Upon
closer inspection, it seems to be some interaction between the latest
versions of the `autolist.nvim` plugin, and the `vim-markdown` and
`vim-polyglot` plugins.

The `vim-markdown` and `vim-polyglot` plugins break the list item
addition in the same way, making me suspect that it's the same fancy
list handling that they try to add (which conflicts with
`autolist.nvim`), so without modifying their code (or configuring them
in some way that I'm not aware of), they both need to be disabled.

I inspected my use of `vim-markdown` and I couldn't find a good reason
to keep it. More than that: I've had issues with the Markdown folding
code causing the folding of the document to flicker and prevent
navigation, causing a rather annoying experience. It could be
`vim-polyglot` causing this, but I'm not sure. Either way, the long list
of features of `vim-markdown` don't interest me, since I already get
syntax highlighting from elsewhere.

Remove the `vim-markdown` plugin, and disable `markdown` for
`vim-polyglot`. Additionally disable `text` for `vim-polyglot`, as the
same broken list handling gets applied, when not much more other code is
loaded for the `text` filetype, making it not worth the load (especially
at the cost of the list handling being broken).