Neovim: remove neomake
I'm not entirely sure if this is something that I use or not, so I'll
presume that it's not the case. I've configured it to handle Elixir, but
it's only covering `credo` and `dogma`, which are sources of warnings
that I already get covered by LSP (either through ElixirLS or by sources
for `null-ls`).