Neovim: add back vim-markdown
In 30e792f7, `vim-markdown` was removed, due me thinking that it would
not work well with `autolist.nvim`. After disabling the plugin, however,
I noticed that there were certain features of `vim-markdown` that I was
missing when having dropped the plugin for just basic syntax
highlighting. The main feature that I realized was missing was the
language specific syntax highlighting for code fences, something that's
very useful for markdown with a lot of code embedding.

I got a tip from the creator of `autolist.nvim` that one could configure
`vim-markdown` to disable it's feature for handling the addition of list
bullets and indentation of the list items. Turning this off does indeed
allow me to use the plugin again.