Neovim: replace vim-speeddating with dial.nvim
While `vim-speeddating` has served me extremely well with no issues over
the years (is it a decade by now?), but this `dial.nvim` plugin seems to
be doing more or less the same thing, but with the added bonus of
having the ability to cycle through a list of words, such as toggling
between `true`/`false`.

I might end up going back if the plugin does not live up to the standard
set by `vim-speeddating`, but I want to try it out.