Neovim: add winbar, only one statusline, and remove command line
With Neovim 0.8, using the `winbar` option makes it easier to get an
idea of which window is which. The `airline` plugin is not very clear
for inactive windows, so `winbar` works better. Set `winbar` to show the
file name and whether the file is modified.

Additionally, `airline` for very narrow windows loses a lot of
information, information that is kept if one has one statusline
(`laststatus=3`) instead.

Finally, `cmdheight=0` gets me back one line of screen space.