Neovim: remove list recalibration for in Insert mode
It seems like the idea of the recalibration feature is that when one is
working in a list, pasting a line will cause numeric list items to have
their indices recalibrated. Currently, this does not work.

For the "p" mapping, my custom `colemak.vim` mapping gets messed up, and
the recalibration happens even when one is not working with a linewise
copy-paste, i.e. the recalibration happens for single words as well
(normally not what I want).

For the "dd" mapping, it seems like dot-repeating gets broken, which is
unacceptable.

Remove the feature completely.

Note that the configuration key name needs to be changed, as the plugin
has had an update which changed/broke the configuration.