Neovim: replace `context.vim` with `nvim-treesitter-context`
While `context.vim` seems to work marginally better than
`nvim-treesitter-context` in a couple of cases (e.g. HEEx files), the
fact that `context.vim` seems to be completely ignorant of the existence
of Neovim's `winbar` option (AFAIK this option is not available in Vim).
This results in `context.vim` overwriting `winbar` whenever there's
a context to be shown, which is acceptable, but very annoying.

Last time I tried out `nvim-treesitter-context` was in 1ad477b, and back
then I hoped that the plugin would improve over time. It seems like it
has, because it works now with nothing but initialization configuration.

Since I like to see my `winbar`, switch `context.vim` for
`nvim-treesitter-context`.