Neovim: set lualine theme
Using the `papercolor` theme seems to be coloring `lualine` as well. The
`papercolor` colors for `lualine` are not bad, but I found the contrast
to be a bit lacking. `powerline_dark` looks nice, and it also gives me
some nostalgia, from back when I was using Powerline a bunch of years
ago.