Neovim: unlock LuaSnip version
The default recommended configuration for LuaSnip included a line for
restricting the version of the plugin. This version caused Sync attempts
to result in an error for the plugin, referencing the version string.

Remove the line. I'm guessing that it can't be that bad to simply get
the latest commit from the repo.