Neovim: replace undotree with telescope-undo
I've had a fork of `mbbill/undotree` for a long time, only due to me
once upon a time thinking that this was the best way to change the
mappings of the plugin. Maybe at that time, this was the only way of
doing it, but I'm not sure.

Either way, I don't want to have forks of plugins, and I'll take the
opportunity to move to a Telescope based plugin, to see if I like it
more.