Neovim: track changes for lazy-lock.json
It's recommended track changes for `lazy-lock.json`, to ensure that the
Neovim plugins are of the same version on all the different machines.

It might be that I'll find it OK that the machines have divergent plugin
versions, since I generally just want to have the latest version at all
times. I might stop tracking this file in the future.