Neovim: lock version of autolist.nvim
The brand new v3.0 version of `autolist.nvim` is a big refactor, which
unfortunately resulted in the plugin freezing the Neovim instance if
recalculation of the bullet point indentation levels is triggered (a
fundamental feature of the plugin).

Until the issue is addressed, use the latest tag that preceded `v3.0`,
i.e. `v2.3`.