Neovim: add cutlass.nvim, for differentiating between cut and delete
This is definitely an experimental addition, but I've generally sided
with those who think that Vim conflating the notion of deleting and
cutting/copying is an annoyance.

A typical example would be when one would want to delete a line (e.g. an
empty one) but retain whatever one has yanked/deleted earlier. This
would necessitate moving whatever one has in register to a named
register, while in every other editor this would not be an issue to even
consider: one would just delete the line.

The behaviour of Vim is so deeply ingrained now that I suspect this will
be an annoyance for quite a while, but I do believe that if this does
not collide with some other plugin, then this is a more sensible for the
editor to work.