Neovim: load vim-abolish eagerly
If one loads `vim-abolish` lazily, Neovim will complain about the lines
in `.config/nvim/after/plugin/abolish.vim` invoking an unknown command,
i.e. the not-yet-defined `:Abolish`.

So far this has not been a problem, as it seems like the lazy loading of
`vim-abolish` happens before the part of Neovim's startup when
`.config/nvim/after/` files are read (whenever that is, especially in
relation to when `lazy.nvim` loads stuff).

To ensure that no such problems occur, e.g. if one sets specific
mappings for the coerce functionality, then the plugin needs to be
eagerly loaded.