Neovim: disable "coerce" functionality of vim-abolish
While the coercion functionality of `vim-abolish` has generally worked
well for me, I'd like to set some different mappings for it, and it
seems like a bit of a hassle to make that work. Since I don't use this
functionality that much, I'll disable it for now.

I'll look around for a Lua-based Neovim plugin for achieving the same
result.