Neovim: replace NERDTree with Neotree
Since I'm moving in the general direction of replacing various Vimscript
plugins with Lua based plugins, NERDTree is next in line.

I've also noticed that NERDTree is currently looking for a new
maintainer. I don't think this necessarily means that the plugin is
falling apart, or something like that, but I'll simply consider this as
an additional reason to switch plugin.