Neovim: add which-key
For quite some time I've had the problem that I forget what things
I have mapped, what the mappings are, and what some of the mappings do.
This plugin should help with all of these. I've been meaning to install
it for quite some time.

The transition to registering all of the various mappings with which-key
is going to take a while. I transitioned some things in this patch, e.g.
the LSP mappings.

As a part of the plugin configuration, I disable a bunch of operator
presets before calling `setup(opts)`. This is done because I use my
`colemak.nvim` plugin, where there are some mappings (e.g. `y` replacing
`w`), which get a slight delay added to them. This would make sense with
default Neovim mappings (i.e. it's OK for a 300ms pause after `y`), but
with my altered mappings (navigating words with `y` instead), the delay
becomes irritating. The setting of operators to `nil` is done to prevent
all default configuration of these mappings.

which-key covers the same features as vim-peekaboo, so remove the
latter plugin.