Zsh: change package update functions for linux-lts
I've decided I'll try using linux-lts for a while, instead of linux-zen.

This means that my package updating scripts can be simplified, as
I might no longer need the extra linux installation step, which is
required for linux-zen because linux-zen-headers is not there in time
for the compilation of the kernel.