Neovim: replace project.nvim with persisted.nvim
I've recently spent some time looking at the available plugins for
working director, workspace, and session management. Out of the plugins
that have a decent amount of stars on GitHub (at least 3-digit amount),
it seems like there are a couple of categories of plugins:

1. Plugins that manage sessions. Auto-saving, auto-loading, different
   sessions for different directories, etc.
2. Jumping to different working directories, mostly referred to as
   "projects". This tracks previously visited project root directories
   (commonly identified by `.git`-folder presence), and allows one to
   easily jump to those directories, wherever one is currently located.
3. Mangers of workspaces, which seems to be a concept that has some
   overlap with the LSP "workspace" concept, but I'm not entirely sure
   about that. Basically, it's a variant of above mentioned "project"
   idea, except that it's one directory upward, with projects being one
   level below from a "workspace".
4. Some kind of combination of the points above.

The first thing I found was that trying to combine a project management
plugin with a session plugin didn't work out well, since the session
plugin didn't really keep up with the CWD changes that were made by the
project plugin.

Secondly, the workspace/project/session plugin [that I considered][1] is
currently in some right-before-1.0-situation, which makes me think that
I'll just was for a while before trying that one out.

Finally, I came to realize that I've always managed the notion of where
I'm starting Neovim/Vim from through the shell, and very rarely used
`lcd` or `cd`. I've also got a variety of tools for jumping around
between projects on the command line, so I'm thinking that a robust
auto-session manager should be enough for satisfying the main need, i.e.
restoring the editor to the state which it was when I last worked with
the particular project. Jumping between projects sounds nice, but it's
not as important as a powerful session handler.

  [1]: https://github.com/gnikdroy/projections.nvim

Considering all of this:

1. Remove `project.nvim`
2. Add `persisted.nvim`
3. Remove `vim-obsession`, as the functionality of this plugin is
   totally covered by `persisted.nvim`.