Neovim: disable which-key preset info for window mappings
I'm not sure why, but the preset mappings for windows (mappings starting
with `<C-w>`) are doing something that causes various freezes of Neovim
after certain `<C-w>` mappings. The one that pushed me to disable the
presets was that `<C-w><C-w>` and `<C-w>w` both just hang the editor,
which is unacceptable.