Neovim: fix setting of "gf" mapping for obsidian.nvim
Paraphrasing from the comment added above the code change:

The default configuration of `obsidian.nvim` only has one mapping: "gf",
for navigation of markdown links. This is functionality that I want, but
the mapping can't be set the normal way (by setting it through
`mappings` key in the `opts`), due to an interaction between
`obsidian.nvim` and `which-key.nvim`.

For `which-key.nvim` to work for various normal-mode mappings that
*start* with "g", it needs to set normal-mode "g" to opening its menu
for "g". Apparently `obsidian.nvim` does not like that, as it seems to
want a completely clear mapping of both normal-mode "g" and "gf"
(presuming that one wants to set "gf" as the mapping).

As a solution to this, set the `mappings` key to an empty table. This
allows setting the mapping without issue in the `config` plugin setup,
after calling the `setup` function with the opts (it sets up
a conditional passthrough). This is the way it currently works, it's
just that the recent versions of `obsidian.nvim` need the empty
`mappings` key to work.